Clean before ci-release
I was bitten by broken cross-release for Scala.js 0.6 and 1.0 because an
`sbt clean` is needed between the two. It's a good practice to always
clean before releasing so I think it's OK to make this change. If a
build relies on running side-effectful operations before `ci-release`
then they can override `CI_CLEAN` to a no-op such as `version`.

```
CI_CLEAN=version sbt side-effect ci-release
```